"Updated homepage layout and added social media icons
This commit includes several updates to the homepage.php file. The changes include:

Declared global variables $bearsamppLang, $bearsamppCore, and $bearsamppHomepage at the beginning of the file.
Added a language attribute to the HTML tag for better accessibility and SEO.
Changed the favicon path from '/bearsampp.ico' to '/icons/app.ico'.
Updated the CSS classes for the navbar to improve alignment and spacing.
Added Discord and Facebook icons with respective links to the navbar.
Updated the Github and Donation icons' source paths and added alt attributes for better accessibility.
These changes aim to improve the user interface, accessibility, and overall user experience of the homepage."